Fixed an issue found by FL; with global nodes the reordering is 'weird'.
Added new argument to define global nodes:

      control.inla = list(global.node.nnbs = m)

where node i is defined as global if nnbs[i] >= m. This option is
additionally to the relative to n-1, global.node.factor, so in total,
a node 'i' is defined as global if

nnbs[i] >=  min( global.node.nnbs, global.node.factor * (n-1))

where n is the size of the graph.

In this commit there has been a lot of work on the 'graph' handling in
INLA, and, unfortunately, these changes where mixed with the above
onces (will I ever learn...). I'll write about these changes soon, but
they need more checking.